+++ packages/docs/src/stories/components/Button.stories.tsx
@@ -1,5 +1,6 @@
+import { Box } from "@/components/ui/box";
+import { Button, ButtonProps } from "@/components/ui/button";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Button, ButtonProps } from "@unimake-ui/react";
 import { MoveRight } from "lucide-react";
 
 const meta = {
@@ -9,23 +10,29 @@ const meta = {
 		Button
 	},
 
+	decorators: Story => <Box>{Story()}</Box>,
+
 	args: {
-		$variant: "primary",
+		variant: "primary",
 		children: "Enviar",
-		$size: "md",
-		disabled: false
+		size: "md",
+		disabled: false,
+		asChild: false,
+		loading: false
 	},
 
 	argTypes: {
-		$variant: {
-			options: ["primary", "secondary", "tertiary", "danger", "safe", "ghost", "link"],
+		variant: {
+			options: ["primary", "secondary", "tertiary", "danger", "safe", "link"],
 			control: { type: "inline-radio" }
 		},
 
-		$size: {
+		size: {
 			options: ["sm", "md", "lg", "full"],
 			control: { type: "inline-radio" }
 		},
+
+		loading: { control: "boolean" },
 
 		disabled: { control: "boolean" }
 	},
@@ -39,49 +46,42 @@ export const Primary: StoryObj<ButtonProps> = {};
 
 export const Secondary: StoryObj<ButtonProps> = {
 	args: {
-		$variant: "secondary",
+		variant: "secondary",
 		children: "Adicionar"
 	}
 };
 
 export const Tertiary: StoryObj<ButtonProps> = {
 	args: {
-		$variant: "tertiary",
+		variant: "tertiary",
 		children: "Cancelar"
 	}
 };
 
 export const Danger: StoryObj<ButtonProps> = {
 	args: {
-		$variant: "danger",
+		variant: "danger",
 		children: "Confirmar"
 	}
 };
 
 export const Safe: StoryObj<ButtonProps> = {
 	args: {
-		$variant: "safe",
+		variant: "safe",
 		children: "Salvar"
-	}
-};
-
-export const Ghost: StoryObj<ButtonProps> = {
-	args: {
-		$variant: "ghost",
-		children: "Ghost"
 	}
 };
 
 export const Link: StoryObj<ButtonProps> = {
 	args: {
-		$variant: "link",
+		variant: "link",
 		children: "Link"
 	}
 };
 
 export const Small: StoryObj<ButtonProps> = {
 	args: {
-		$size: "sm"
+		size: "sm"
 	}
 };
 
@@ -89,7 +89,7 @@ export const WithIcon: StoryObj<ButtonProps> = {
 	args: {
 		children: (
 			<>
-				Próximo passo
+				<span>Próximo</span>
 				<MoveRight />
 			</>
 		)