+++ packages/docs/src/stories/components/Checkbox.stories.tsx
@@ -1,5 +1,7 @@
+import { Box } from "@/components/ui/box";
+import { Checkbox, CheckboxProps } from "@/components/ui/checkbox";
+import { Text } from "@/components/ui/text";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Box, Checkbox, CheckboxProps, Text } from "@unimake-ui/react";
 
 const meta = {
 	title: "Form/Checkbox",
@@ -8,14 +10,27 @@ const meta = {
 		Checkbox
 	},
 
-	args: {},
+	args: {
+		size: "md",
+		disabled: false
+	},
+
+	argTypes: {
+		size: {
+			control: "inline-radio",
+			options: ["sm", "md"]
+		},
+		disabled: {
+			control: "boolean"
+		}
+	},
 
 	decorators: [
 		Story => {
 			return (
 				<Box as="label" style={{ display: "flex", flexDirection: "row", alignItems: "center", gap: "8px" }}>
 					{Story()}
-					<Text $size="sm">Encerrado?</Text>
+					<Text size="sm">Encerrado?</Text>
 				</Box>
 			);
 		}
@@ -30,7 +45,7 @@ export const Primary: StoryObj<CheckboxProps> = {};
 
 export const Small: StoryObj<CheckboxProps> = {
 	args: {
-		$size: "sm"
+		size: "sm"
 	}
 };