+++ packages/docs/src/stories/components/TextArea.stories.tsx
@@ -1,5 +1,7 @@
+import { Box } from "@/components/ui/box";
+import { Text } from "@/components/ui/text";
+import { TextArea, TextAreaProps } from "@/components/ui/text-area";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Box, Text, TextArea, TextAreaProps } from "@unimake-ui/react";
 
 const meta = {
 	title: "Form/Text Area",
@@ -10,7 +12,7 @@ const meta = {
 
 	args: {
 		disabled: false,
-		$hasError: false,
+		hasError: false,
 		placeholder: ""
 	},
 
@@ -23,7 +25,7 @@ const meta = {
 			type: "string"
 		},
 
-		$hasError: {
+		hasError: {
 			type: "boolean"
 		}
 	},
@@ -32,7 +34,7 @@ const meta = {
 		Story => {
 			return (
 				<Box as="label" style={{ display: "flex", flexDirection: "column", gap: "4px" }}>
-					<Text $size="sm">Descrição</Text>
+					<Text size="sm">Descrição</Text>
 					{Story()}
 				</Box>
 			);
@@ -54,7 +56,7 @@ export const Placeholder: StoryObj<TextAreaProps> = {
 
 export const WithError: StoryObj<TextAreaProps> = {
 	args: {
-		$hasError: true
+		hasError: true
 	}
 };