+++ packages/tokens/src/line-heights.ts
@@ -1,6 +1,7 @@
 export const lineHeights = {
-  shorter: '125%',
-  short: '140%',
-  base: '160%',
-  tall: '180%',
+	full: "100%",
+	shorter: "125%",
+	short: "140%",
+	base: "160%",
+	tall: "180%"
 } as const;