+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unimake-ui/ts-config
 
+## 1.0.3
+
+### Patch Changes
+
+- Add @unimake-ui/tailwindcss package and mdify some tokens for better adapting with components
+
 ## 1.0.2
 
 ### Patch Changes